Option with for-comprehension tests are moved to another package
